refactor: main controller 분리
