--- src/test/java/ggs/srr/security/AuthenticationHolderTest.java
@@ -1,6 +1,0 @@
-package ggs.srr.security;
-
-
-
-class AuthenticationHolderTest {
-}
